Create country phone list
