+++ src/main/webapp/app/modules/new-profile/create-new-profile/pages/determine-responsibility-and-follow-up/determineResponsibilityAndFollowUp.page.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from "react-router";
 import { ButtonComponent, DropDownComponent, InputComponent } from "@eachawy/frontend-library";
 import { useForm } from "react-hook-form";
 import { useAppSelector } from "app/config/store";
-import { countryCode } from "app/shared/util/date-utils";
+import PhoneNumberComponent from "app/shared/components/phoneNumber.Component/phoneNumber.Component";
 
 const DetermineResponsibilityAndFollowUpPage = () => {
 
@@ -30,26 +30,6 @@ const DetermineResponsibilityAndFollowUpPage = () => {
 
   const saveAndCloseFn = () => {
     navigate("/create-new-profile");
-  };
-
-  const selectedCountryCodeTemplate = (option) => {
-    if (option) {
-      return (
-        <div className="countryCodeTemplate">
-          <span className={`flag-icon flag-icon-${option.code.toLowerCase()} `}></span>
-          <div>{option.name}</div>
-        </div>
-      );
-    }
-  };
-
-  const countryCodeOptionTemplate = (option) => {
-    return (
-      <div className="countryCodeTemplate">
-        <span className={`flag-icon flag-icon-${option.code.toLowerCase()} `}></span>
-        <div>{option.name}</div>
-      </div>
-    );
   };
 
   return (
@@ -79,40 +59,9 @@ const DetermineResponsibilityAndFollowUpPage = () => {
             onChange={(e) => setValue("email", e.target.value)}
             rules={{ required: 'يجب ادخال البريد الالكتروني' }}
           />
-          <div className="phoneNoDiv col-md-6">
-            <label>رقم الهاتف <span>*</span></label>
-            <div>
-              <DropDownComponent
-                id="DetermineResponsibilityAndFollowUpPhoneNumber-id"
-                name="DRAFCountryCode"
-                register={register}
-                watch={watch}
-                setValueMethod={setValue}
-                options={countryCode}
-                optionLabel={`name.${lang}`}
-                errors={errors}
-                onChange={(e) => setValue("DRAFCountryCode", e.value)}
-                setValue={countryCode[0]}
-                valueTemplate={selectedCountryCodeTemplate}
-                itemTemplate={countryCodeOptionTemplate}
-              />
-              <InputComponent
-                id="DRAFPhoneNo-id"
-                type="text"
-                name="DRAFPhoneNo"
-                placeholder="مثال: 1234567"
-                register={register}
-                errors={errors}
-                setValueMethod={setValue}
-                watch={watch}
-                onChange={(e) => {
-                  const numericValue = e.target.value.replace(/[^0-9]/g, "");
-                  setValue("DRAFPhoneNo", numericValue);
-                }}
-                value={watch("DRAFPhoneNo")}
-              />
-            </div>
-          </div>
+
+          <PhoneNumberComponent register={register} errors={errors} watch={watch} setValue={setValue} />
+
         </div>
         <div className="actionBtns">
           <ButtonComponent Class={'BtnCancel'} onClick={saveAndCloseFn}>{translate("createNewProfile.saveAndClose")}</ButtonComponent>
